+++ src/main/java/tipitapi/drawmytoday/oauth/dto/UserProfile.java
@@ -1,4 +1,4 @@
-package tipitapi.drawmytoday.common.security.oauth2.dto;
+package tipitapi.drawmytoday.oauth.dto;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import lombok.AllArgsConstructor;
