+++ src/main/java/tipitapi/drawmytoday/oauth/entity/AppleProperties.java
@@ -1,4 +1,4 @@
-package tipitapi.drawmytoday.common.security.oauth2.entity;
+package tipitapi.drawmytoday.oauth.entity;
 
 import java.util.Date;
 import lombok.Getter;
