+++ src/main/java/tipitapi/drawmytoday/oauth/service/AppleOAuthService.java
@@ -1,4 +1,4 @@
-package tipitapi.drawmytoday.common.security.oauth2.service;
+package tipitapi.drawmytoday.oauth.service;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
@@ -18,11 +18,11 @@ import org.springframework.util.MultiValueMap;
 import org.springframework.util.StringUtils;
 import org.springframework.web.client.RestTemplate;
 import tipitapi.drawmytoday.common.security.jwt.JwtTokenProvider;
-import tipitapi.drawmytoday.common.security.oauth2.dto.AppleIdToken;
-import tipitapi.drawmytoday.common.security.oauth2.dto.RequestAppleLogin;
-import tipitapi.drawmytoday.common.security.oauth2.dto.ResponseAccessToken;
-import tipitapi.drawmytoday.common.security.oauth2.dto.ResponseJwtToken;
-import tipitapi.drawmytoday.common.security.oauth2.entity.AppleProperties;
+import tipitapi.drawmytoday.oauth.dto.AppleIdToken;
+import tipitapi.drawmytoday.oauth.dto.RequestAppleLogin;
+import tipitapi.drawmytoday.oauth.dto.ResponseAccessToken;
+import tipitapi.drawmytoday.oauth.dto.ResponseJwtToken;
+import tipitapi.drawmytoday.oauth.entity.AppleProperties;
 import tipitapi.drawmytoday.user.domain.Auth;
 import tipitapi.drawmytoday.user.domain.OAuthType;
 import tipitapi.drawmytoday.user.domain.User;
